middleware_project/requirements.txt
pika
python-dotenv
flask
requests
